backend/pyproject.toml
[project]
name = "agent-prototype"
version = "0.1.0"
description = "Multi-tenant conversational assistant PoC"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.12.0",
    "redis>=5.0.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-postgres>=0.0.1",
    "pgvector>=0.2.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*"]

[tool.black]
line-length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
